[FE] RENAME: social account link card 디렉토리명 변경 #1739
