+++ frontend/src/Cabinet/components/Card/SocialAccountLinkCard/SnsConnectionCard.tsx
