+++ frontend/src/Cabinet/components/Card/SocialAccountLinkCard/SnsConnectionCardModal.tsx
