+++ frontend/src/Cabinet/components/Card/SocialAccountLinkCard/SocialAccountLink.container.tsx
@@ -2,8 +2,8 @@ import { HttpStatusCode } from "axios";
 import { useState } from "react";
 import { useRecoilState } from "recoil";
 import { userState } from "@/Cabinet/recoil/atoms";
-import SnsConnectionCard from "@/Cabinet/components/Card/SnsConnectionCard/SnsConnectionCard";
-import SnsConnectionCardModal from "@/Cabinet/components/Card/SnsConnectionCard/SnsConnectionCardModal";
+import SnsConnectionCard from "@/Cabinet/components/Card/SocialAccountLinkCard/SnsConnectionCard";
+import SnsConnectionCardModal from "@/Cabinet/components/Card/SocialAccountLinkCard/SnsConnectionCardModal";
 import {
   TOAuthProvider,
   ftProvider,
